Added slime trail ability

+++ justraces-api/src/main/java/justjabka/JustRaces/Abilities/Generic/DurationAbility.java
@@ -3,7 +3,7 @@ package justjabka.JustRaces.Abilities.Generic;
 import org.bukkit.entity.Player;
 
 public interface DurationAbility extends ResettableAbility {
-    int getDurationTicks();
+    long getDurationTicks();
     default void onExpire(Player player) {
         resetState(player, Reason.ABILITY_END);
     }
