+++ justraces-showcase/src/main/java/justjabka/JustRacesShowcase/Configs/Ability/EcdysisAbilityConfig.java
@@ -4,7 +4,7 @@ import org.bukkit.configuration.file.FileConfiguration;
 
 public class EcdysisAbilityConfig {
     public final long cooldown;
-    public final int effectDuration;
+    public final long effectDuration;
     public final float explosionPowerNormal;
     public final float explosionPowerSuicide;
     public final float durabilityPercentPerUse;
@@ -13,7 +13,7 @@ public class EcdysisAbilityConfig {
 
     public EcdysisAbilityConfig(FileConfiguration config) {
         this.cooldown = config.getLong("cooldown", 9) * 20;
-        this.effectDuration = config.getInt("effect-duration", 11) * 20;
+        this.effectDuration = config.getLong("effect-duration", 11) * 20;
         this.explosionPowerNormal = (float) config.getDouble("explosion-power.normal", 6);
         this.explosionPowerSuicide = (float) config.getDouble("explosion-power.suicide", 7);
         this.durabilityPercentPerUse = (float) config.getDouble("durability-percent-per-use", 0.25);
